refactor: restructure project folders
- data 패키지 내의 usecase 구현체를 domain 모듈로 이동했습니다
- API 서버로 부터 받게 되는 폴더를 entity 로 변경하고 로컬에서 데이터 전달 시
사용되는 domain 모듈 내의 model 을 dto 로 변경하였습니다
