+++ app/src/main/java/campus/tech/kakao/map/data/remote/entity/MetaEntity.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.data.remote.dto
+package campus.tech.kakao.map.data.remote.entity
 
 import com.google.gson.annotations.SerializedName
