+++ app/src/main/java/campus/tech/kakao/map/data/remote/entity/SearchResponse.kt
@@ -1,4 +1,4 @@
-package campus.tech.kakao.map.data.remote.dto
+package campus.tech.kakao.map.data.remote.entity
 
 data class SearchResponse(
     val meta: MetaEntity,
