+++ app/src/main/java/campus/tech/kakao/map/di/UsecaseModule.kt
@@ -1,12 +1,11 @@
 package campus.tech.kakao.map.di
 
-import campus.tech.kakao.map.data.usecase.GetLastPlaceUseCaseImpl
-import campus.tech.kakao.map.data.usecase.GetSearchHistoryUseCaseImpl
-import campus.tech.kakao.map.data.usecase.GetSearchPlacesUseCaseImpl
-import campus.tech.kakao.map.data.usecase.RemoveSearchQueryUseCaseImpl
-import campus.tech.kakao.map.data.usecase.SaveLastPlaceUseCaseImpl
-import campus.tech.kakao.map.data.usecase.SaveSearchQueryUseCaseImpl
-import campus.tech.kakao.map.domain.repository.PlaceRepository
+import campus.tech.kakao.map.domain.usecaseImpl.GetLastPlaceUseCaseImpl
+import campus.tech.kakao.map.domain.usecaseImpl.GetSearchHistoryUseCaseImpl
+import campus.tech.kakao.map.domain.usecaseImpl.GetSearchPlacesUseCaseImpl
+import campus.tech.kakao.map.domain.usecaseImpl.RemoveSearchQueryUseCaseImpl
+import campus.tech.kakao.map.domain.usecaseImpl.SaveLastPlaceUseCaseImpl
+import campus.tech.kakao.map.domain.usecaseImpl.SaveSearchQueryUseCaseImpl
 import campus.tech.kakao.map.domain.usecase.GetLastPlaceUseCase
 import campus.tech.kakao.map.domain.usecase.GetSearchHistoryUseCase
 import campus.tech.kakao.map.domain.usecase.GetSearchPlacesUseCase
@@ -15,10 +14,8 @@ import campus.tech.kakao.map.domain.usecase.SaveLastPlaceUseCase
 import campus.tech.kakao.map.domain.usecase.SaveSearchQueryUseCase
 import dagger.Binds
 import dagger.Module
-import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
 
 @InstallIn(SingletonComponent::class)
 @Module
