+++ app/src/main/java/campus/tech/kakao/map/domain/usecase/GetSearchPlacesUseCase.kt
@@ -1,6 +1,6 @@
 package campus.tech.kakao.map.domain.usecase
 
-import campus.tech.kakao.map.domain.model.PlaceVO
+import campus.tech.kakao.map.domain.dto.PlaceVO
 
 interface GetSearchPlacesUseCase {
     suspend operator fun invoke(query: String): List<PlaceVO>?
