+++ app/src/main/java/campus/tech/kakao/map/domain/usecase/SaveSearchQueryUseCase.kt
@@ -1,6 +1,6 @@
 package campus.tech.kakao.map.domain.usecase
 
-import campus.tech.kakao.map.domain.model.PlaceVO
+import campus.tech.kakao.map.domain.dto.PlaceVO
 
 interface SaveSearchQueryUseCase {
     suspend operator fun invoke(place: PlaceVO)
