+++ app/src/main/java/campus/tech/kakao/map/domain/usecaseImpl/GetLastPlaceUseCaseImpl.kt
@@ -1,7 +1,6 @@
-package campus.tech.kakao.map.data.usecase
+package campus.tech.kakao.map.domain.usecaseImpl
 
-import android.util.Log
-import campus.tech.kakao.map.domain.model.PlaceVO
+import campus.tech.kakao.map.domain.dto.PlaceVO
 import campus.tech.kakao.map.domain.repository.PlaceRepository
 import campus.tech.kakao.map.domain.usecase.GetLastPlaceUseCase
 import javax.inject.Inject
