+++ app/src/main/java/campus/tech/kakao/map/domain/usecaseImpl/GetSearchHistoryUseCaseImpl.kt
@@ -1,8 +1,7 @@
-package campus.tech.kakao.map.data.usecase
+package campus.tech.kakao.map.domain.usecaseImpl
 
 import campus.tech.kakao.map.domain.repository.PlaceRepository
 import campus.tech.kakao.map.domain.usecase.GetSearchHistoryUseCase
-import campus.tech.kakao.map.domain.usecase.GetSearchPlacesUseCase
 import javax.inject.Inject
 
 class GetSearchHistoryUseCaseImpl @Inject constructor(private val placeRepository: PlaceRepository) :
