+++ app/src/main/java/campus/tech/kakao/map/domain/usecaseImpl/SaveLastPlaceUseCaseImpl.kt
@@ -1,8 +1,7 @@
-package campus.tech.kakao.map.data.usecase
+package campus.tech.kakao.map.domain.usecaseImpl
 
-import campus.tech.kakao.map.domain.model.PlaceVO
+import campus.tech.kakao.map.domain.dto.PlaceVO
 import campus.tech.kakao.map.domain.repository.PlaceRepository
-import campus.tech.kakao.map.domain.usecase.GetLastPlaceUseCase
 import campus.tech.kakao.map.domain.usecase.SaveLastPlaceUseCase
 import javax.inject.Inject
