+++ app/src/main/java/campus/tech/kakao/map/domain/usecaseImpl/SaveSearchQueryUseCaseImpl.kt
@@ -1,6 +1,6 @@
-package campus.tech.kakao.map.data.usecase
+package campus.tech.kakao.map.domain.usecaseImpl
 
-import campus.tech.kakao.map.domain.model.PlaceVO
+import campus.tech.kakao.map.domain.dto.PlaceVO
 import campus.tech.kakao.map.domain.repository.PlaceRepository
 import campus.tech.kakao.map.domain.usecase.SaveSearchQueryUseCase
 import javax.inject.Inject
