+++ app/src/main/java/campus/tech/kakao/map/presentation/MapViewModel.kt
@@ -1,11 +1,10 @@
 package campus.tech.kakao.map.presentation
 
-import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import campus.tech.kakao.map.domain.model.PlaceVO
+import campus.tech.kakao.map.domain.dto.PlaceVO
 import campus.tech.kakao.map.domain.usecase.GetLastPlaceUseCase
 import campus.tech.kakao.map.domain.usecase.SaveLastPlaceUseCase
 import dagger.hilt.android.lifecycle.HiltViewModel
